feat: WIP - extract sub components to ui folder
