--- src/features/domain-settings/DomainSettingsForm/DomainSettingsForm.types.ts
@@ -1,16 +1,0 @@
-import { Step } from '@zero-tech/zui/components';
-
-export const steps: Step[] = [
-	{
-		id: 'details',
-		title: 'Details',
-	},
-	{
-		id: 'confirm',
-		title: 'Confirm',
-	},
-	{
-		id: 'complete',
-		title: 'Complete',
-	},
-];
