+++ src/features/ui/Media/index.ts
@@ -1,0 +1,1 @@
+export * from './Media';
