+++ src/features/ui/TextArea/TextArea.module.scss
@@ -1,0 +1,33 @@
+.TextAreaContainer {
+	display: flex;
+	flex-direction: column;
+	gap: 0.25rem;
+
+	.TextArea {
+		background: none;
+		color: white;
+		border: none;
+		font-size: 0.875rem;
+		line-height: 1.05625rem;
+
+		outline: none;
+		min-height: 2.5rem;
+		padding: 0.25rem 0.5rem;
+
+		border-radius: 0.5rem;
+		border: 1px solid #404040;
+		resize: vertical;
+		cursor: text;
+
+		transition: border-color 0.1s ease;
+
+		&:disabled {
+			background: #1c1c1f;
+			resize: none;
+
+			&:hover {
+				cursor: not-allowed;
+			}
+		}
+	}
+}
