+++ src/features/ui/index.ts
@@ -6,7 +6,9 @@ export * from './FormInputs';
 export * from './FormTextContent';
 export * from './FormUserBalance';
 export * from './IconButton';
+export * from './Media';
 export * from './OptionLabel';
 export * from './StatsList';
 export * from './Switch';
 export * from './TableCard';
+export * from './TextArea';
